WIP: Refactor avatar modules
