+++ src/lib/fitting/FittingAvatar.js
@@ -1,7 +1,83 @@
+import * as THREE from "@/lib/threejs/three";
+import { readByteArray } from "@/lib/clo/file/KeyValueMapReader";
+import FittingSkinControllerManager from "@/lib/fitting/FittingSkinControllerManager";
+import ResizableBody from "@/lib/fitting/FittingResizableBody";
+import Accessory from "@/lib/fitting/FittingAccessory";
+import FittingAccessory from "@/lib/fitting/FittingAccessory";
+
 export default class FittingAvatar {
   constructor(container, zrest) {
     this.container = container;
+    this.zrest = zrest;
+
+    // NOTE: This is named "list" but actually a map. This name according to the CLO SW.
+    // this.listSkinController = new Map();
+
+    // NOTE: This map is created by parsing the listSkinController.
+    //       Keys are the name of the SkinController. After creating this map, listSkinController is deallocated from memory.
+    // this.mapSkinController = new Map();
+
+    // this.avatarContainer = null;
+    // this.accessoryContainer = null;
+
+    this.accessory = null;
+
+    // this.scManager = new FittingSkinControllerManager(this.zrest);
+    this.resizableBody = null;
+    this.scManager = null;
+
+    this.bodyVertexIndex = [];
+    this.bodyVertexPos = [];
   }
+
+  init() {
+    if (!this.zrest) return;
+
+    this.avatarContainer = new THREE.Object3D();
+    this.avatarContainer.name = "fittingAvatarContainer";
+    this.container.add(this.avatarContainer);
+
+    // this.container.add(this.accessory.getThreeJSContainer());
+    // this.accessory
+
+    // this.accessoryContainer.name = "fittingAccessoryContainer";
+    // this.scene.add(this.accessoryContainer);
+
+    const avatarGeometry = new Map(
+      this.zrest.zProperty.rootMap.get("mapGeometry")
+    );
+    const listSkinController = this.loadGeometry({
+      mapGeometry: avatarGeometry,
+    });
+    this.setAvatarInfo(listSkinController);
+
+    this.scManager = new FittingSkinControllerManager(this.zrest);
+    this.scManager.init(this.zrest);
+
+    this.accessory = new FittingAccessory(listSkinController, this.scManager);
+    this.accessory.attachThreeJSContainer(this.container);
+    this.accessory.putBodyVertexInfo(this.bodyVertexPos, this.bodyVertexIndex);
+
+    // this.mapSkinController = this.convertListSCtoMap(listSkinController);
+    // console.warn("mapSkinController");
+    // console.warn(this.mapSkinController)
+  }
+
+  // Init resizable body and accessory
+  initResizableBodyWithAcc(avatarSizingInfoObj) {
+    this.resizableBody = new ResizableBody({
+      gender: 0,
+      mapBaseMesh: avatarSizingInfoObj.mapBaseMesh,
+      convertingMatData: avatarSizingInfoObj.convertingMatData,
+      mapHeightWeightTo5Sizes: avatarSizingInfoObj.mapHeightWeightTo5Sizes,
+      mapAccessoryMesh: avatarSizingInfoObj.mapAccessoryMesh,
+      scManager: this.scManager,
+    });
+
+    // TODO: Move this module to the proper position.
+    this.accessory.putMeshInfo(avatarSizingInfoObj.mapAccessoryMesh);
+  }
+
   async resize({
     height,
     weight,
@@ -57,6 +133,17 @@ export default class FittingAvatar {
     }
   }
 
+  resizeAccessory() {
+    if (this.resizeAccessory) {
+      this.accessory.putBodyVertexInfo(
+        this.bodyVertexPos,
+        this.bodyVertexIndex
+      );
+      this.accessory.resize();
+    }
+    // else console.warn("Can't access accessory");
+  }
+
   getAvatarURL({ id: avatarId, skinType: avatarSkinType }) {
     this.avatarId = avatarId;
 
@@ -65,5 +152,74 @@ export default class FittingAvatar {
     const avtURL = this.avtRootPath + "/" + zrestFileName;
     console.log(avtURL);
     return avtURL;
+  }
+
+  loadGeometry({ mapGeometry: mapGeometry }) {
+    this.extractController(mapGeometry);
+    // this.convertListSCtoMap(this.listSkinController);
+    return this.listSkinController;
+  }
+
+  setAvatarInfo(listSkinController) {
+    const bodySkinController = this.findBodySkinController(listSkinController);
+    // console.log("bodySkin is");
+    // console.log(bodySkinController);
+    this.bodySkinController = bodySkinController;
+
+    const mapMesh = bodySkinController.get("mapMesh");
+    const meshIndex = readByteArray("Uint", mapMesh.get("baIndex"));
+    const meshPosition = readByteArray("Float", mapMesh.get("baPosition"));
+    this.bodyVertexIndex = meshIndex;
+    this.bodyVertexPos = meshPosition;
+
+    // TODO: Set this function correctly
+    // this.garment.setBody(this.bodyVertexPos, this.bodyVertexIndex);
+
+    //this.accessory.init
+
+    // NOTE: For test only
+    // this.buildMeshUsingMapMesh(mapMesh);
+  }
+
+  // TODO: Refactor this module
+  extractController(mapInput) {
+    const shouldRecursive = (element) => {
+      return (
+        element instanceof Map &&
+        element.has("listChildrenTransformer3D") &&
+        element.get("listChildrenTransformer3D") != null
+      );
+    };
+
+    if (shouldRecursive(mapInput)) {
+      this.extractController(mapInput.get("listChildrenTransformer3D"));
+    } else {
+      mapInput.forEach((inputElement) => {
+        if (shouldRecursive(inputElement)) {
+          return this.extractController(inputElement);
+        }
+        if (inputElement.has("listSkinController")) {
+          console.log(inputElement);
+          console.log(inputElement.get("listSkinController"));
+          this.listSkinController = inputElement.get("listSkinController");
+        }
+      });
+    }
+  }
+
+  findBodySkinController(listSkinController) {
+    let largestSC = null;
+    let largestLength = 0;
+    listSkinController.forEach((sc) => {
+      if (sc.has("baInitPosition")) {
+        const length = sc.get("baInitPosition").byteLength;
+        if (largestLength < length) {
+          largestLength = length;
+          largestSC = sc;
+        }
+      }
+    });
+
+    return largestSC;
   }
 }
